# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Codespace X",
        "title": "Codespace X REST API",
        "contact": {
            "name": "Codespace Indonesia",
            "url": "https://codespace.id",
            "email": "mail@codespace.id"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/v1/auth/exchange-token": {
            "post": {
                "description": "Exchange Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Exchange Token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-Service-Auth-Token",
                        "name": "X-Service-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "payload exchange token",
                        "name": "body-payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ExchangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ExchangeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/otp/request": {
            "post": {
                "description": "OTP Request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "OTP Request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-Service-Auth-Token",
                        "name": "X-Service-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "payload otp request",
                        "name": "body-payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OtpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/otp/resend": {
            "post": {
                "description": "OTP Resend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "OTP Resend",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-Service-Auth-Token",
                        "name": "X-Service-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "payload otp request resend",
                        "name": "body-payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OtpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/otp/validate": {
            "post": {
                "description": "OTP Validate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "OTP Validate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-Service-Auth-Token",
                        "name": "X-Service-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "payload otp request",
                        "name": "body-payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OtpValidateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/phone/verify": {
            "post": {
                "description": "Used when register before hit register button",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Phone Verification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-Service-Auth-Token",
                        "name": "X-Service-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "payload otp request resend",
                        "name": "body-payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OtpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ExchangeRequest": {
            "type": "object",
            "required": [
                "firebase_id_token",
                "phone_number"
            ],
            "properties": {
                "firebase_id_token": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "dto.ExchangeResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "dto.OtpRequest": {
            "type": "object",
            "required": [
                "phone_number"
            ],
            "properties": {
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "dto.OtpValidateRequest": {
            "type": "object",
            "required": [
                "otp",
                "phone_number"
            ],
            "properties": {
                "otp": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
        "pkg.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/pkg.MetaResponse"
                }
            }
        },
        "pkg.MetaResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                }
            }
        }
    }
}